Moved schema to examples, added a generic schema, and a comma
This incorporates the kernel of what Simeon discussed in WECG and
recommended in
https://github.com/w3c/webextensions/pull/585#issuecomment-2077956982.
It's not meant to replace that comment, but only to introduce it here
for clarity, while allowing Simeon to update it as needed once this PR
is merged.

Added a fuzzy matching section again even though someone suggested it
be removed earlier. It was introduced in 6e6af62a8503, but removed in a
subsequent "Commit suggestion" from somoene on GitHub. It's worth
calling out here at some point, so it might as well be in this PR,
which is already approved but still awaiting someone with the ability
to go ahead and merge it.

Some of these changes may foster discussion, which is completely
understandable.