Use latest Ubuntu version for GitHub deploy workflow
Our deploy workflow was failing because it was using an older version of
Ubuntu (which in turn had an older version of Python). Update our YML
configuration to use the latest version.

I considered pinning to a newer version but this matches both our other
workflows and the w3c/spec-prod repo [1] so seems like a better choice.

[1] https://github.com/w3c/spec-prod/blob/main/.github/workflows/docs.yml